+++ src/backend/resolvers/mutations/product/create.js
@@ -1,6 +1,30 @@
-const createProduct = (parent, { input: { _id, ...body } }, { pulseCoreDb }, info) => {
-  return pulseCoreDb.collection('products').insertOne(body)
-    .then(res => res.ops[0])
+const axios = require('axios')
+const { v4: uuid } = require('uuid')
+
+const createProduct = async (
+  parent,
+  { input: { _id, ...body } },
+  { pulseCoreDb },
+  info
+) => {
+  // ! Vega Op
+  const vegaId = uuid()
+
+  const vegaProductInput = {
+    id: vegaId,
+    brand_name: body.nameBrand,
+    generic_name: body.nameGeneric,
+  }
+
+  await axios.post('products/', vegaProductInput).catch((e) => {
+    throw new Error(JSON.stringify(e.response.data))
+  })
+
+  // ! Mongo Op
+  return pulseCoreDb
+    .collection('products')
+    .insertOne({ ...body, uuid: vegaId })
+    .then((res) => res.ops[0])
 }
 
 module.exports = createProduct