+++ CHANGELOG.md
@@ -12,7 +12,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - **Fixed** for any bug fixes.
 - **Security** in case of vulnerabilities.
 
-## [Unreleased] - Started: 2-26-21
+## [Unreleased] - Started: 3-9-21
+
+## [4.7.0] - Started: 2-26-21
 
 ### Added
 - `Perception Tool/Vega Migration`: